+++ client/setup.py
@@ -1,10 +1,24 @@
 from setuptools import setup, find_packages
 
 setup(
-    name="mlmm",
+    name="Maisie",
     version="0.1",
     packages=find_packages(),
     include_package_data=True,
-    install_requires=["Click"],
-    entry_points={"console_scripts": ["mlmm = mlmm.interface:cli"]},
+    install_requires=[
+        "Click",
+        "gitpython",
+        "requests",
+        "coloredlogs",
+        "pyyaml",
+        "terminaltables",
+        "ago",
+        "python-dateutil",
+    ],
+    entry_points={
+        "console_scripts": [
+            "mai = maisie.interface:cli",
+            "maisie = maisie.interface:cli",
+        ]
+    },
 )
